requirements.txt
matplotlib
pillow
opencv-python
numpy
cv2